requirement.txt
streamlit
pytesseract
Pillow
pdf2image
langchain
langchain-core
langchain-ollama
openai-whisper
torch
sounddevice
scipy
gTTS
playsound==1.2.2
numpy
webrtcvad
soundfile
ffmpeg-python
setuptools<81
pydub
python-dotenv
edge-tts